Move module-name resolution logic from `infer.rs` to `module_name.rs`
